requirements.txt
# Core modules (mandatory)
jupyterlite-core==0.1.2
jupyterlab~=4.0.5

# Python kernel (optional)
jupyterlite-pyodide-kernel==0.1.1

# JavaScript kernel (optional)
jupyterlite-javascript-kernel==0.1.2

# Language support (optional)
jupyterlab-language-pack-fr-FR
jupyterlab-language-pack-zh-CN

# SQLite kernel (optional)
jupyterlite-xeus-sqlite==0.2.1
# P5 kernel (optional)
jupyterlite-p5-kernel==0.1.1
# Lua kernel (optional)
jupyterlite-xeus-lua==0.3.3

# Python: ipywidget library for Jupyter notebooks (optional)
ipywidgets>=8.0.0,<9
# Python: ipyevents library for Jupyter notebooks (optional)
ipyevents>=2.0.1
# Python: interative Matplotlib library for Jupyter notebooks (optional)
ipympl>=0.8.2
# Python: ipycanvas library for Jupyter notebooks (optional)
ipycanvas>=0.9.1
# Python: ipyleaflet library for Jupyter notebooks (optional)
ipyleaflet

# Python: plotting libraries (optional)
plotly>=5,<6
bqplot

# Custom requirements
tskit
msprime
tsinfer
tsdate
demesdraw
seaborn
jupyterquiz
jupyterlab-filesystem-access
networkx
pyyaml
drawsvg
